# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Travel Agency API
  version: 1.0.0
  description: |
    旅行社在线跟团游系统的正式接口契约。PRD 定义业务，本文件定义可执行的 HTTP 契约。
    当前后端脚手架不构成兼容依据；前端 Mock、正式后端和契约测试均以本文件为准。
...
servers:
  - url: /api
    description: 当前站点 API
security:
  - bearerAuth: []
...
paths:
...
  /admin/hotels:
    get:
      tags: [Admin Resources]
      operationId: listAdminHotels
      summary: 分页查询酒店资料
      x-roles: [STAFF, ADMIN]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - name: keyword
          in: query
          schema: { type: string, maxLength: 100 }
      responses:
        '200':
          description: 酒店分页结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HotelPageEnvelope' }
    post:
      tags: [Admin Resources]
      operationId: createHotel
      summary: 创建酒店资料
      x-roles: [STAFF, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HotelUpsertRequest' }
      responses:
        '201':
          description: 酒店创建成功
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HotelEnvelope' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /admin/hotels/{hotelId}:
    parameters:
      - { $ref: '#/components/parameters/HotelId' }
    put:
      tags: [Admin Resources]
      operationId: updateHotel
      summary: 修改酒店资料
      x-roles: [STAFF, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HotelUpsertRequest' }
      responses:
        '200':
          description: 修改后的酒店
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HotelEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [Admin Resources]
      operationId: deleteHotel
      summary: 删除未被行程引用的酒店资料
      x-roles: [STAFF, ADMIN]
      responses:
        '204': { description: 删除成功 }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    Location:
      description: 新建资源的访问路径
      schema: { type: string, format: uri-reference }
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    HotelId: { name: hotelId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
  responses:
...
    NotFound:
      description: 资源不存在或对当前用户不可见
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            code: RESOURCE_NOT_FOUND
            message: 目标资源不存在
            data: null
            errors: []
            traceId: 550e8400-e29b-41d4-a716-446655440000
    Conflict:
      description: 重复提交、并发或资源状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            code: RESOURCE_STATE_CONFLICT
            message: 当前资源状态不允许此操作
            data: null
            errors: []
            traceId: 550e8400-e29b-41d4-a716-446655440000
    ValidationFailed:
      description: 字段语义校验或业务规则校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            code: VALIDATION_ERROR
            message: 请求参数校验失败
            data: null
            errors:
              - field: contactPhone
                code: INVALID_FORMAT
                message: 手机号格式不正确
            traceId: 550e8400-e29b-41d4-a716-446655440000
...
  schemas:
    Id:
      type: string
      pattern: '^[1-9][0-9]*$'
      example: '1001'
...
    DateTime:
      type: string
      format: date-time
      example: '2026-09-07T14:30:00+08:00'
    Longitude:
      type: number
      minimum: -180
      maximum: 180
      example: 102.8328910
    Latitude:
      type: number
      minimum: -90
      maximum: 90
      example: 24.8800950
    AccountStatus:
      type: string
      enum: [ACTIVE, DISABLED]
      example: ACTIVE
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: [string, 'null']
          example: contactPhone
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          example: INVALID_FORMAT
        message:
          type: string
          example: 手机号格式不正确
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, errors, traceId]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        message: { type: string }
        data: { type: 'null' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
        traceId: { type: string, format: uuid }
    SuccessEnvelope:
      type: object
      required: [code, message, data, errors, traceId]
      properties:
        code: { type: string, const: OK, example: OK }
        message: { type: string, const: success, example: success }
        data: {}
        errors:
          type: array
          maxItems: 0
        traceId: { type: string, format: uuid, example: 550e8400-e29b-41d4-a716-446655440000 }
    PageFields:
      type: object
      required: [page, size, total, totalPages]
      properties:
        page: { type: integer, minimum: 1, example: 1 }
        size: { type: integer, minimum: 1, maximum: 100, example: 20 }
        total: { type: integer, minimum: 0, example: 1 }
        totalPages: { type: integer, minimum: 0, example: 1 }
...
    Hotel:
      type: object
      additionalProperties: false
      required: [id, name, status, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name: { type: string, example: 昆明测试酒店 }
        address: { type: [string, 'null'], example: 云南省昆明市测试路 1 号 }
        contactPhone: { type: [string, 'null'], example: '087112345678' }
        longitude:
          anyOf:
            - { $ref: '#/components/schemas/Longitude' }
            - { type: 'null' }
        latitude:
          anyOf:
            - { $ref: '#/components/schemas/Latitude' }
            - { type: 'null' }
        intro: { type: [string, 'null'], example: 仅用于课程演示的酒店资料。 }
        dataSource: { type: [string, 'null'], example: 团队测试数据 }
        status: { $ref: '#/components/schemas/AccountStatus' }
        createdAt: { $ref: '#/components/schemas/DateTime' }
        updatedAt: { $ref: '#/components/schemas/DateTime' }
    HotelUpsertRequest:
      type: object
      additionalProperties: false
      required: [name, dataSource]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        address: { type: [string, 'null'], maxLength: 255 }
        contactPhone: { type: [string, 'null'], maxLength: 20 }
        longitude:
          anyOf:
            - { $ref: '#/components/schemas/Longitude' }
            - { type: 'null' }
        latitude:
          anyOf:
            - { $ref: '#/components/schemas/Latitude' }
            - { type: 'null' }
        intro: { type: [string, 'null'], maxLength: 10000 }
        dataSource: { type: string, minLength: 1, maxLength: 500 }
        status: { $ref: '#/components/schemas/AccountStatus' }
...
    HotelPage:
      allOf:
        - { $ref: '#/components/schemas/PageFields' }
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Hotel' }
...
    HotelEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Hotel' }
    HotelPageEnvelope:
      allOf:
        - { $ref: '#/components/schemas/SuccessEnvelope' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/HotelPage' }
...